Refactor: DTO에 Record 클래스 적용 - e.printStackTrace() 대체 - setter 사용 지양
