+++ build.gradle
@@ -12,7 +12,7 @@ checkstyle {
     toolVersion = "8.24"
 }
 
-group = 'com.example'
+group = 'java.dev'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '17'
 
@@ -42,7 +42,7 @@ dependencies {
 
     // GitHub API For Java
     implementation group: 'org.kohsuke', name: 'github-api', version: '1.313'
-    
+
     // JWT 구현을 위한 jsonwebtoken 의존성 추가
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
