--- src/main/java/com/example/donmit/DonMitApplication.java
@@ -1,13 +1,0 @@
-package com.example.donmit;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DonMitApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DonMitApplication.class, args);
-	}
-
-}
