--- src/main/java/com/example/donmit/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.example.donmit.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
