--- src/main/java/com/example/donmit/config/SecurityConfig.java
@@ -1,53 +1,0 @@
-package com.example.donmit.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import com.example.donmit.model.util.JwtTokenProvider;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-	private final JwtTokenProvider jwtTokenProvider;
-
-	private static final String[] AUTH_WHITELIST = {
-		"/", "/login/oauth2/**", "/v3/api-docs/**", "/swagger-ui/**"
-	};
-
-	@Bean
-	public AuthenticationManager authenticationManager(
-		AuthenticationConfiguration authenticationConfiguration
-	) throws Exception {
-		return authenticationConfiguration.getAuthenticationManager();
-	}
-
-	@Bean
-	protected SecurityFilterChain config(HttpSecurity http) throws Exception {
-		return http
-			.httpBasic().disable()
-			.formLogin().disable()
-			.csrf().disable()
-			.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-			.and()
-
-			.authorizeHttpRequests(auth -> auth
-				.requestMatchers(AUTH_WHITELIST).permitAll()
-				.anyRequest().authenticated()
-			)
-
-			.addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class)
-			.build();
-
-	}
-}
