--- src/main/java/com/example/donmit/model/dto/res/OauthToken.java
@@ -1,18 +1,0 @@
-package com.example.donmit.model.dto.res;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class OauthToken {
-
-	private String access_token;
-	private String scope;
-	private String token_type;
-
-}
