+++ src/main/java/dev/donmit/donmitapi/DonMitApplication.java
@@ -1,0 +1,12 @@
+package dev.donmit.donmitapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DonMitApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(DonMitApplication.class, args);
+	}
+
+}
