+++ src/main/java/dev/donmit/donmitapi/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package dev.donmit.donmitapi.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
