+++ src/main/java/dev/donmit/donmitapi/config/JwtAuthenticationFilter.java
@@ -1,0 +1,40 @@
+package dev.donmit.donmitapi.config;
+
+import java.io.IOException;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.GenericFilterBean;
+
+import dev.donmit.donmitapi.model.util.JwtTokenProvider;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+
+/*
+ * 클라이언트 요청 시 JWT 인증을 위한 커스텀 필터로,
+ * UsernamePasswordAuthenticationFilter 이전에 실행되는 필터
+ */
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends GenericFilterBean {
+	private final JwtTokenProvider jwtTokenProvider;
+
+	@Override
+	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws
+		IOException, ServletException {
+		// 헤더로부터 JWT를 받아온다.
+		String token = jwtTokenProvider.resolveToken((HttpServletRequest)request);
+
+		// 유효한 토큰인지 확인한다.
+		if (token != null && jwtTokenProvider.validateToken(token)) {
+			Authentication authentication = jwtTokenProvider.getAuthentication(token);
+			
+			// SecurityContext에 Authentication 객체를 저장한다.
+			SecurityContextHolder.getContext().setAuthentication(authentication);
+		}
+		chain.doFilter(request, response);
+	}
+}
