+++ src/main/java/dev/donmit/donmitapi/model/dto/res/OauthToken.java
@@ -1,0 +1,4 @@
+package dev.donmit.donmitapi.model.dto.res;
+
+public record OauthToken(String access_token, String scope, String token_type) {
+}
