+++ src/main/java/dev/donmit/donmitapi/model/dto/res/TokenResDto.java
@@ -1,0 +1,8 @@
+package dev.donmit.donmitapi.model.dto.res;
+
+import lombok.Builder;
+
+@Builder
+public record TokenResDto(String grantType, Long accessTokenExpireDate, String accessToken, String refreshToken) {
+}
+
