+++ src/main/java/dev/donmit/donmitapi/model/util/BaseEntity.java
@@ -1,0 +1,23 @@
+package dev.donmit.donmitapi.model.util;
+
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class BaseEntity {
+	@CreatedDate
+	@Column(updatable = false)
+	private LocalDateTime createdDate;
+	@LastModifiedDate
+	private LocalDateTime modifiedDate;
+}
