+++ src/main/resources/application.yml
@@ -1,0 +1,20 @@
+spring:
+  datasource:
+    url: jdbc:mysql://localhost:3306/donmit
+    username: # database username
+    password: # database password
+
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.MySQLDialect
+
+
+secretKey: # secretKey
+
+client_id: # GitHub OAuth App Clined ID
+client_secret: # GitHub OAuth App Clined Secret
+
