--- src/test/java/com/example/donmit/DonMitApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.donmit;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DonMitApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
