+++ src/test/java/dev/donmit/donmitapi/DonMitApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.donmit.donmitapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DonMitApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
